Move Result from run to webserver package
Result is no longer produced by the Runner to be sent to the Webserver
via a channel. Instead, the Webserver is tasked with regularly updating
the Result from the kubernetes state.
