+++ webserver/result.go
@@ -1,4 +1,4 @@
-package run
+package webserver
 
 import (
 	"fmt"
